Load manifest information for Illumina Infinium Omni5-4 SNP bead chips

This dataset is a combination of the following two datasets from https://support.illumina.com/array/array_kits/humanomni5-4-beadchip-kit/downloads.html:

load_manifest()

Format

A data frame with 4,327,108 rows (probes) and 18 variables:

Name

Name of probe

IlmnID

Illumina ID of probe

RsID

rs number of probe

Chr

Chromosome

AddressA_ID

Address A of probe

AddressB_ID

Address B of probe

SNP

The possible bases of the SNP

IlmnStrand

TOP/BOT/PLUS/MINUS

SourceStrand

TOP/BOT/PLUS/MINUS

TopGenomicSeqSBE

The content of the .[.]. part of TopGenomicSeq (not included), including the flanking left and right parts

TopGenomicSeqSBE_Left

The left flanking of TopGenomicSeqSBE

TopGenomicSeqSBE_Right

The right flanking of TopGenomicSeqSBE

BeadSetID

Bead set ID

Exp_Clusters

3 or 2

RefStrand

+/-

ProbeType

II (only address A) or I (address A and B)

SNPType

Signals whether SNP has alleles on same color channel (ambiguous) or not (unambiguous), or potentially is an INDEL (insertion/deletion): AMBiguous for AT-SNPs ([A/T] or [T/A]) and CG-SNPs; INDEL for INDELs; UNAMBiguous for the rest (e.g. AC-SNPs)

Exclude

Some probes should be excluded (cf. above), this indicate those

Source

https://support.illumina.com/array/array_kits/humanomni5-4-beadchip-kit/downloads.html

Details

InfiniumOmni5-4v1-2_A2.csv

Infinium Omni5-4 v1.2 Product Files -> Infinium Omni5-4 v1.2 Manifest File (CSV Format - GRCh38)

InfiniumOmni5-4v1-2_A1_b144_rsids.txt

Infinium Omni5-4 v1.2 Support Files -> Infinium Omni5-4 v1.2 Loci Name to rsID Conversion File

Note that the following probes have Exclude = FALSE (all others have Exclude = TRUE) because they are ambiguous (SNP is C/G), but they are a Type-II probe (i.e. do not have an AddressB_ID), so the bases cannot be distinguish as both C and G are detected by the green channel:

Name rs28362918 (Illumina ID rs28362918-131_T_R_1908372611)

Name rs28897688 (Illumina ID rs28897688-131_T_R_1908379219)